# Conflicts:
#	bluedragon-distribution-core/pom.xml
#	bluedragon-distribution-core/src/main/java/com/jd/bluedragon/distribution/station/dao/WorkStationGridDao.java
#	bluedragon-distribution-core/src/main/java/com/jd/bluedragon/distribution/station/dao/impl/WorkStationGridDaoImpl.java
#	bluedragon-distribution-core/src/main/java/com/jd/bluedragon/distribution/station/service/UserSignRecordService.java
#	bluedragon-distribution-core/src/main/java/com/jd/bluedragon/distribution/station/service/WorkStationGridService.java
#	bluedragon-distribution-core/src/main/java/com/jd/bluedragon/distribution/station/service/impl/UserSignRecordServiceImpl.java
#	bluedragon-distribution-core/src/main/java/com/jd/bluedragon/distribution/station/service/impl/WorkStationGridServiceImpl.java
#	bluedragon-distribution-core/src/main/resources/mappers/mysql/station/WorkStationGridMapper.xml
#	bluedragon-distribution-sdk/pom.xml
#	bluedragon-distribution-utils/src/main/java/com/jd/bluedragon/dms/utils/BusinessUtil.java

+++ bluedragon-distribution-core/src/main/java/com/jd/bluedragon/distribution/station/service/impl/WorkStationGridServiceImpl.java
@@ -4,16 +4,22 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 
 import org.apache.commons.collections.CollectionUtils;
-import org.springframework.beans.BeanUtils;
+import org.apache.commons.lang.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
 
+import com.jd.bluedragon.Constants;
 import com.jd.bluedragon.core.base.BaseMajorManager;
 import com.jd.bluedragon.core.objectid.IGenerateObjectId;
 import com.jd.bluedragon.distribution.api.response.base.Result;
+import com.jd.bluedragon.distribution.position.domain.PositionRecord;
+import com.jd.bluedragon.distribution.position.service.PositionRecordService;
 import com.jd.bluedragon.distribution.station.dao.WorkStationGridDao;
 import com.jd.bluedragon.distribution.station.domain.WorkStation;
 import com.jd.bluedragon.distribution.station.domain.WorkStationGrid;
@@ -53,20 +59,43 @@ public class WorkStationGridServiceImpl implements WorkStationGridService {
 	private BaseMajorManager baseMajorManager;
 	@Autowired
 	private IGenerateObjectId genObjectId;
+
+	@Autowired
+	private PositionRecordService positionRecordService;
 	/**
 	 * 插入一条数据
 	 * @param insertData
 	 * @return
 	 */
+	@Transactional(value = "main_undiv", propagation = Propagation.REQUIRED, rollbackFor = Exception.class)
 	public Result<Boolean> insert(WorkStationGrid insertData){
 		Result<Boolean> result = checkAndFillBeforeAdd(insertData);
 		if(!result.isSuccess()) {
 			return result;
 		}
-		insertData.setBusinessKey(DmsConstants.CODE_PREFIX_WORK_STATION_GRID.concat(StringHelper.padZero(this.genObjectId.getObjectId(WorkStationGrid.class.getName()),11)));
+		insertData.setBusinessKey(generalBusinessKey());
 		result.setData(workStationGridDao.insert(insertData) == 1);
+		// 添加岗位记录
+		addPosition(insertData);
 		return result;
 	 }
+
+	private String generalBusinessKey() {
+		return DmsConstants.CODE_PREFIX_WORK_STATION_GRID.concat(StringHelper.padZero(this.genObjectId.getObjectId(WorkStationGrid.class.getName()),11));
+	}
+
+	private void addPosition(WorkStationGrid insertData) {
+		PositionRecord record = new PositionRecord();
+		record.setSiteCode(insertData.getSiteCode());
+		record.setRefGridKey(insertData.getBusinessKey());
+		record.setCreateUser(insertData.getCreateUser());
+		record.setUpdateUser(insertData.getUpdateUser());
+		Result<Integer> result = positionRecordService.insertPosition(record);
+		if(result == null || !Objects.equals(result.getData(), Constants.YN_YES)){
+			throw new RuntimeException("添加岗位数据失败，岗位对应的业务主键是:" + insertData.getBusinessKey());
+	}
+	}
+
 	/**
 	 * 校验并填充数据add
 	 * @param insertData
@@ -176,9 +205,23 @@ public class WorkStationGridServiceImpl implements WorkStationGridService {
 	 * @param deleteData
 	 * @return
 	 */
+	@Transactional(value = "main_undiv", propagation = Propagation.REQUIRED, rollbackFor = Exception.class)
 	public Result<Boolean> deleteById(WorkStationGrid deleteData){
 		Result<Boolean> result = Result.success();
+		Result<WorkStationGrid> queryResult = queryById(deleteData.getId());
+		if(queryResult.getData() == null || StringUtils.isEmpty(queryResult.getData().getBusinessKey())){
+			throw new RuntimeException("根据id:" + deleteData.getId() + "未查询到数据!");
+		}
 		result.setData(workStationGridDao.deleteById(deleteData) == 1);
+		// 同步删除岗位记录
+		String businessKey = queryResult.getData().getBusinessKey();
+		PositionRecord positionRecord = new PositionRecord();
+		positionRecord.setRefGridKey(businessKey);
+		positionRecord.setUpdateUser(deleteData.getUpdateUser());
+		Result<Boolean> deletePositionResult = positionRecordService.deleteByBusinessKey(positionRecord);
+		if(Objects.equals(deletePositionResult.getData(), false)){
+			throw new RuntimeException("根据businessKey:" + businessKey + "删除岗位数据失败!");
+		}
 		return result;
 	 }
 	/**
@@ -188,7 +231,7 @@ public class WorkStationGridServiceImpl implements WorkStationGridService {
 	 */
 	public Result<WorkStationGrid> queryById(Long id){
 		Result<WorkStationGrid> result = Result.success();
-		result.setData(toWorkStationGrid(workStationGridDao.queryById(id)));
+		result.setData(workStationGridDao.queryById(id));
 		return result;
 	 }
 	/**
@@ -233,26 +276,15 @@ public class WorkStationGridServiceImpl implements WorkStationGridService {
 		
 		return result;
 	 }
-	/**
-	 * 对象转换成vo
-	 * @param data
-	 * @return
-	 */
-	public WorkStationGrid toWorkStationGrid(WorkStationGrid data){
-		if(data == null) {
-			return null;
-		}
-		WorkStationGrid vo = new WorkStationGrid();
-		BeanUtils.copyProperties(data, vo);
-		//特殊字段设置
-		return vo;
-	 }
+
 	@Override
 	public Result<List<WorkStationGrid>> queryAllGridBySiteCode(WorkStationGridQuery query) {
 		Result<List<WorkStationGrid>> result = Result.success();
 		result.setData(workStationGridDao.queryAllGridBySiteCode(query));
 		return result;
 	}
+
+	@Transactional(value = "main_undiv", propagation = Propagation.REQUIRED, rollbackFor = Exception.class)
 	@Override
 	public Result<Boolean> importDatas(List<WorkStationGrid> dataList) {
 		Result<Boolean> result = checkAndFillImportDatas(dataList);
@@ -266,15 +298,28 @@ public class WorkStationGridServiceImpl implements WorkStationGridService {
 				oldData.setUpdateUser(data.getCreateUser());
 				oldData.setUpdateUserName(data.getCreateUserName());
 				oldData.setUpdateTime(data.getCreateTime());
-				workStationGridDao.deleteById(oldData);
+				if(!Objects.equals(workStationGridDao.deleteById(oldData), Constants.YN_YES)){
+					throw  new RuntimeException("根据id:" + oldData.getId() + "删除数据失败!");
+				}
 				data.setBusinessKey(oldData.getBusinessKey());
 			}else {
-				data.setBusinessKey(DmsConstants.CODE_PREFIX_WORK_STATION_GRID.concat(StringHelper.padZero(this.genObjectId.getObjectId(WorkStationGrid.class.getName()),11)));
+				data.setBusinessKey(generalBusinessKey());
 			}			
-			workStationGridDao.insert(data);
+			if(!Objects.equals(workStationGridDao.insert(data), Constants.YN_YES)){
+				throw  new RuntimeException("新增businessKey为:" + data.getBusinessKey() + "的数据失败");
+			}
+			// 同步处理岗位
+			syncDealPosition(oldData, data);
 		}
 		return result;
 	}
+
+	private void syncDealPosition(WorkStationGrid oldData, WorkStationGrid newData) {
+		if(oldData == null){
+		addPosition(newData);
+	}
+	}
+
 	/**
 	 * 校验并填充导入数据
 	 * @param dataList
@@ -319,7 +364,7 @@ public class WorkStationGridServiceImpl implements WorkStationGridService {
 	@Override
 	public Result<WorkStationGrid> queryByBusinessKey(WorkStationGrid data) {
 		Result<WorkStationGrid> result = Result.success();
-		result.setData(toWorkStationGrid(workStationGridDao.queryByBusinessKey(data)));
+		result.setData(workStationGridDao.queryByBusinessKey(data));
 		return result;
 	}
 	@Override
@@ -363,5 +408,12 @@ public class WorkStationGridServiceImpl implements WorkStationGridService {
 	@Override
 	public boolean hasGridData(String stationKey) {
 		return workStationGridDao.queryCountByRefStationKey(stationKey) > 0;
+	}
+
+	@Override
+	public Result<WorkStationGrid> queryByGridKey(WorkStationGridQuery workStationGridQuery) {
+		Result<WorkStationGrid> result = Result.success();
+		result.setData(workStationGridDao.queryByGridKey(workStationGridQuery));
+		return result;
 	}
 }
